refactor. convert some page into TS
