+++ packages/namu/src/helper/index.js
@@ -1,3 +1,4 @@
 export * from "./debounce.js";
 export * from "./dom.js";
-export * from "./storage.js";
+export * from "./storage.ts";
+export * from "./api.ts";
