--- packages/namu/src/helper/storage.js
@@ -1,7 +1,0 @@
-export function setLocal(key, value) {
-  localStorage.setItem(key, value);
-}
-
-export function getLocal(key) {
-  return localStorage.getItem(key);
-}
